+++ src/types.ts
@@ -2,6 +2,7 @@ export type ChecksumAddress = string;
 export type HexEncodedBytes = string;
 export type Base64EncodedBytes = string;
 
-export interface Configuration {
-  porterUri: string;
+export enum ChainId {
+  POLYGON = 137,
+  MUMBAI = 80001,
 }